add an additional unit test that could cause users to issue tokens without holding the tokens at low quantity issues
